fix: updated type for SystemCheckNavigation
